+++ .pico-testrc.json
@@ -1,0 +1,5 @@
+{
+    "tests": [
+        "test/tests/*.test.js"
+    ]
+}
