+++ test/framework/.eslintrc.json
@@ -5,5 +5,9 @@
     "env": {
         "node": true,
         "es6": true
+    },
+    "rules": {
+        "global-require": 0,
+        "no-empty": 0
     }
 }
