+++ test/framework/pico-test.js
@@ -5,10 +5,23 @@ const pti = require("puppeteer-to-istanbul");
 const http = require("http");
 const fs = require("fs").promises;
 const path = require("path");
+const glob = require("glob");
 
-const PATH = "test-results/";
-const PORT = 7171;
-const BASE_URL = `http://localhost:${PORT}/`;
+let config = {
+    outputDir: "test-results/",
+    serverPort: 7171,
+    tests: []
+};
+
+try {
+    Object.assign(config, require(path.resolve(".", ".pico-testrc.json")));
+    if (!Array.isArray(config.tests)) {
+        config.tests = [ config.tests ];
+    }
+} catch (e) {}
+
+const CWD = process.cwd();
+const BASE_URL = `http://localhost:${config.serverPort}/`;
 const FILTER_REGEX = new RegExp(`${BASE_URL}src/`);
 const MIME_TYPES = {
     ".css": "text/css",
@@ -16,13 +29,6 @@ const MIME_TYPES = {
     ".js": "text/javascript",
     ".json": "application/json"
 };
-
-const TESTS = [
-    "test/tests/picogl.test.js",
-    "test/tests/app.test.js",
-    "test/tests/cubemap.test.js",
-    "test/tests/texture.test.js"
-];
 
 const server = http.createServer(async (req, res) => {
     const url = req.url; 
@@ -41,15 +47,21 @@ const server = http.createServer(async (req, res) => {
         res.setHeader("Content-Type", mimeType);
         res.end(content);
     }
-}).listen(PORT);
+}).listen(config.serverPort);
 
 (async () => {
     const template = await fs.readFile("test/framework/index-template.html", "utf8");
-    const index = template.replace("FRAMEWORK_IMPORTS", TESTS.map(t => `import "../${t}";`).join("\n"));
+    const testPaths = [];
+    config.tests.forEach(t => {
+        glob.sync(t).map(f => {
+            const modulePath = path.resolve(".", f).replace(CWD, "");
+            testPaths.push(`import "${modulePath}";`);
+        });
+    });
+    const index = template.replace("FRAMEWORK_IMPORTS", testPaths.join("\n"));
 
     await fs.mkdir("test-results");
     await fs.writeFile("test-results/index.html", index);
-    await fs.copyFile("test/framework/js/pico-test.js", "test-results/pico-test.js");
 
     const browser = await puppeteer.launch();
     const page = await browser.newPage();
@@ -98,5 +110,5 @@ const server = http.createServer(async (req, res) => {
     });
 
     await page.coverage.startJSCoverage();
-    await page.goto(`${BASE_URL}${PATH}`);
+    await page.goto(`${BASE_URL}${config.outputDir}`);
 })();
