+++ test/tests/.eslintrc.json
@@ -6,5 +6,8 @@
     "env": {
         "browser": true,
         "es6": true
+    },
+    "globals": {
+        "picoTest": true
     }
 }
