--- test/tests/index.js
@@ -1,4 +1,0 @@
-import "./picogl.test.js";
-import "./app.test.js";
-import "./cubemap.test.js";
-import "./texture.test.js";
